leds: refactor util functions
