+++ components/leds/include/leds.h
@@ -313,6 +313,8 @@ struct leds_color {
   };
 };
 
+bool leds_color_active (struct leds_color color, enum leds_parameter_type parameter_type);
+
 enum leds_test_mode {
   TEST_MODE_BLACK,
   TEST_MODE_CHASE,
