+++ components/leds/leds.h
@@ -35,6 +35,12 @@ int leds_init(struct leds *leds, const struct leds_options *options);
 /* interface.c */
 int leds_interface_init(union leds_interface_state *interface, const struct leds_protocol_type *protocol_type, const struct leds_options *options);
 
+/* color.c */
+unsigned leds_colors_active (const struct leds_color *colors, unsigned count, enum leds_parameter_type parameter_type);
+
+/* power.c */
+unsigned leds_power_total(const struct leds_color *pixels, unsigned index, unsigned count, enum leds_power_mode power_mode);
+
 /* format.c */
 void leds_set_format_rgb(struct leds *leds, uint8_t *data, size_t len, struct leds_format_params params);
 void leds_set_format_bgr(struct leds *leds, uint8_t *data, size_t len, struct leds_format_params params);
